refactor(pane): Phase 2 — lift Win32 focus-redirect into pane/hwnd.rs
Moves \`PANE_WNDPROCS\`, \`ALLOW_PANE_FOCUS_ONCE\`, and
\`install_pane_focus_redirect\` from \`client.rs\` to \`agentmux-cef/src/pane/hwnd.rs\`.
Pure code motion. No behavior change.

Files
-----

\`agentmux-cef/src/pane/hwnd.rs\` (new, Windows-only via \`#![cfg]\`):
- \`PANE_WNDPROCS\` — map of subclassed HWND -> original WndProc pointer.
- \`ALLOW_PANE_FOCUS_ONCE\` — focus-bypass AtomicBool (now \`pub\` in the
  \`pane\` module instead of \`client\`).
- \`install_pane_focus_redirect\` — the subclass installer + inner
  \`wndproc_hook\` + \`EnumChildWindows\` traversal.
- 3 new \`pane::hwnd::tests\` for the AtomicBool swap/load semantics (the
  one part of this file that's testable without a real HWND).

\`agentmux-cef/src/pane/mod.rs\`:
- \`#[cfg(target_os = "windows")] pub mod hwnd;\` + re-exports of
  \`ALLOW_PANE_FOCUS_ONCE\` and \`install_pane_focus_redirect\`.

\`agentmux-cef/src/client.rs\`:
- ~140 lines removed. Replaced with a three-line pointer comment.

\`agentmux-cef/src/browser_panes.rs\`:
- \`focus()\` switches \`crate::client::ALLOW_PANE_FOCUS_ONCE\` →
  \`crate::pane::ALLOW_PANE_FOCUS_ONCE\`.

Known gap (documented, not fixed here)
-------------------------------------

\`install_pane_focus_redirect\` has zero callers. Per
\`SPEC_BROWSER_PANE_LIFECYCLE.md\` §5 race #5, it should be invoked from
pane \`on_after_created\` and on every \`on_load_end\` (Chromium creates a
new \`Chrome_RenderWidgetHostHWND\` on every navigation). That wiring lands
in Phase 4 of the split when pane-specific CEF callbacks move out of
\`client.rs\` into \`pane/callbacks.rs\` — doing it in Phase 2 would mix
code motion with behavior change.

Testing
-------

\`cargo test\` — 27 tests pass (24 pre-existing + 3 new hwnd tests).
Runtime < 0.01s. No production call-path change.
