+++ agentmux-cef/src/client.rs
@@ -1008,144 +1008,9 @@ static ORIGINAL_WNDPROCS: std::sync::LazyLock<
     std::sync::Mutex<std::collections::HashMap<usize, isize>>,
 > = std::sync::LazyLock::new(|| std::sync::Mutex::new(std::collections::HashMap::new()));
 
-/// Map of pane HWND -> original WndProc for the focus-redirect subclass below.
-#[cfg(target_os = "windows")]
-static PANE_WNDPROCS: std::sync::LazyLock<
-    std::sync::Mutex<std::collections::HashMap<usize, isize>>,
-> = std::sync::LazyLock::new(|| std::sync::Mutex::new(std::collections::HashMap::new()));
-
-/// Subclass a browser pane's HWND so that WM_SETFOCUS is immediately redirected
-/// back to the parent (top-level) window. Without this, Chromium's internal
-/// SetFocus on the pane HWND (on page load, on JS window.focus() calls, etc.)
-/// steals the Windows-level keyboard focus — any subsequent keystrokes go to
-/// the pane's renderer instead of the main window, so terminals, URL bars, and
-/// other inputs in the main UI stop responding.
-///
-/// The pane becomes view-only (user cannot type into google.com's search box),
-/// but the main window keeps working. A later gesture (e.g. Alt+click) can
-/// restore focus to the pane explicitly when the user wants to interact with
-/// the embedded page.
-/// When true, the next WM_SETFOCUS delivered to a subclassed pane HWND is
-/// allowed through instead of being redirected. The `browser_pane_focus` IPC
-/// handler sets this flag before calling SetFocus on the pane, so user-
-/// initiated focus (routed by the frontend's ViewModel.giveFocus()) works
-/// even though the Chromium-internal focus steal on navigation is blocked.
-#[cfg(target_os = "windows")]
-pub static ALLOW_PANE_FOCUS_ONCE: std::sync::atomic::AtomicBool =
-    std::sync::atomic::AtomicBool::new(false);
-
-#[cfg(target_os = "windows")]
-unsafe fn install_pane_focus_redirect(hwnd: *mut std::ffi::c_void) {
-    use std::sync::atomic::Ordering;
-    use windows_sys::Win32::UI::WindowsAndMessaging::{
-        CallWindowProcW, GetParent, SetWindowLongPtrW, GWLP_WNDPROC, WM_SETFOCUS,
-    };
-    use windows_sys::Win32::UI::Input::KeyboardAndMouse::SetFocus;
-
-    unsafe extern "system" fn wndproc_hook(
-        hwnd: *mut std::ffi::c_void,
-        msg: u32,
-        wparam: usize,
-        lparam: isize,
-    ) -> isize {
-        // Diagnostic: surface mouse-wheel and key events so we can tell
-        // whether they reach the pane HWND at all when the user reports
-        // scrolling/typing breakage.
-        const WM_MOUSEWHEEL: u32 = 0x020A;
-        const WM_MOUSEHWHEEL: u32 = 0x020E;
-        const WM_KEYDOWN: u32 = 0x0100;
-        const WM_CHAR: u32 = 0x0102;
-        match msg {
-            WM_MOUSEWHEEL | WM_MOUSEHWHEEL => {
-                tracing::info!("[pane-wndproc] mouse-wheel hwnd={:p} msg=0x{:x}", hwnd, msg);
-            }
-            WM_KEYDOWN | WM_CHAR => {
-                tracing::info!("[pane-wndproc] key msg=0x{:x} wparam={}", msg, wparam);
-            }
-            _ => {}
-        }
-
-        if msg == WM_SETFOCUS {
-            // Intentional focus from the frontend's giveFocus() IPC: honor it
-            // once, then revert to redirect-mode for subsequent events.
-            if ALLOW_PANE_FOCUS_ONCE.swap(false, Ordering::Relaxed) {
-                tracing::info!("[pane-wndproc] WM_SETFOCUS allowed (intentional)");
-                // Fall through to the original WndProc.
-            } else {
-                // Programmatic focus (page load, JS window.focus()): redirect.
-                let parent = GetParent(hwnd);
-                if !parent.is_null() {
-                    SetFocus(parent);
-                }
-                return 0;
-            }
-        }
-
-        let original = PANE_WNDPROCS
-            .lock()
-            .ok()
-            .and_then(|m| m.get(&(hwnd as usize)).copied())
-            .unwrap_or(0);
-        if original != 0 {
-            let proc_fn: unsafe extern "system" fn(
-                *mut std::ffi::c_void, u32, usize, isize,
-            ) -> isize = std::mem::transmute(original);
-            CallWindowProcW(Some(proc_fn), hwnd, msg, wparam, lparam)
-        } else {
-            0
-        }
-    }
-
-    // Subclass the outer HWND — but only once. Re-calling SetWindowLongPtrW
-    // would replace our hook with itself and poison PANE_WNDPROCS.
-    let already_hooked = PANE_WNDPROCS
-        .lock()
-        .ok()
-        .map(|m| m.contains_key(&(hwnd as usize)))
-        .unwrap_or(false);
-    if !already_hooked {
-        let original = SetWindowLongPtrW(hwnd, GWLP_WNDPROC, wndproc_hook as *const () as isize);
-        if original != 0 {
-            if let Ok(mut map) = PANE_WNDPROCS.lock() {
-                map.insert(hwnd as usize, original);
-            }
-            tracing::info!("[pane-subclass] installed focus-redirect WndProc on pane HWND {:p}", hwnd);
-        }
-    }
-
-    // Chromium creates inner HWNDs (widget + render) below the outer HWND.
-    // Mouse input reaches the deepest descendant, so we must walk the whole
-    // tree and subclass every one.
-    unsafe extern "system" fn enum_children(
-        child: *mut std::ffi::c_void,
-        _lparam: isize,
-    ) -> i32 {
-        let already = PANE_WNDPROCS
-            .lock()
-            .ok()
-            .map(|m| m.contains_key(&(child as usize)))
-            .unwrap_or(false);
-        if already {
-            return 1;
-        }
-        let orig = SetWindowLongPtrW(child, GWLP_WNDPROC, wndproc_hook as *const () as isize);
-        if orig != 0 {
-            if let Ok(mut map) = PANE_WNDPROCS.lock() {
-                map.insert(child as usize, orig);
-            }
-            let mut class_buf = [0u16; 64];
-            let n = windows_sys::Win32::UI::WindowsAndMessaging::GetClassNameW(
-                child, class_buf.as_mut_ptr(), class_buf.len() as i32,
-            );
-            let class_name = String::from_utf16_lossy(&class_buf[..n as usize]);
-            tracing::info!("[pane-subclass] subclassed child HWND {:p} class={}", child, class_name);
-        }
-        1 // continue
-    }
-    windows_sys::Win32::UI::WindowsAndMessaging::EnumChildWindows(
-        hwnd, Some(enum_children), 0,
-    );
-}
+// Pane Win32 focus-redirect subclass + ALLOW_PANE_FOCUS_ONCE flag moved to
+// `crate::pane::hwnd` in Phase 2 of the modularization split. See
+// `docs/specs/SPEC_BROWSER_PANE_MODULARIZATION.md`.
 
 /// Install a WndProc hook on a SECONDARY window that handles:
 /// - WM_NCCALCSIZE: returns 0 to eliminate the non-client area (removes the
